feat: docker compose file
Based on the upstream Penumbra setup, but here we add
a postgres container and of course the cuiloa container itself.

So far, it initializes a database and runs a Penumbra node.
It seems to run the Cuiloa app, as well, but the landing
for the app is blank, implying the db connection isn't working.

What works:

  * pd/cometbft join network and fetch blocks
  * init script sets psql values in cometbft config
  * cometbft schema exists in postgres

What doesn't work:

  * cuiloa app shows blank landing page
  * cuiloa app never complains, even if it can't reach db

What I've checked so far:

  * cometbft does not emit errors related to auth
  * postgres tables appear empty, implying no events ingested