ci: unbreak containerfile build
The container build has been broken since merge of [0], but that failure
was easy to miss since we don't run tests on PRs in the cuiloa repo.
Also, the deploy workflow is no longer working, but we still want to
build containers on every merge into main, so the build code is
accessible in testing and deploy environments.

[0] https://github.com/penumbra-zone/cuiloa/pull/132/files