feat: support CA_CERT env var override
When connecting to remote databases that are SSL-required, the app will
sometimes report cert errors. First, I tried to resolve this by pulling
in `ca-certificates` in the container environment, but that wasn't
sufficient. Turns out `node-postgres` in v8.x updated the ssl logic and
that broke setups for many managed databases. As a workaround, one can
now:

  1. set the CA_CERT env var with the string contents of the db's CA
  2. remove the `sslmode=require` from the connection auth starting

The fact that 2 is necessary is a bit scary, but it's caused by the fact
that the connectionString setting clobbers any manual `ssl` opts in the
db config.